requirements.txt
huggingface-hub>=0.19.0
transformers>=4.36.0
torch>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
jsonschema>=4.0.0
rich>=13.0.0
streamlit>=1.29.0
